Removed windows includes from OgrePrerequisites.h. MS's dirty use of 'boolean' and other types causes problems and there is just no need for all OGRE files to include them. Altered files which required windows headers to include them directly. Removed all the special build reduction macros since unnecessary given smaller dependencies. Fixed a problem with the JPEG codec crashing because the struct sizes were incompatible. NEW DEPENDENCY DOWNLOAD REQUIRED, jpeg headers and libraries have been updated to latest. Changed Char to OgreChar since clashes with MS headers.

+++ trunk/OgreMain/include/OgreJPEGCodec.h
@@ -45,7 +45,7 @@ namespace Ogre {
 		/* boolean is actually int from libJPEG's point of view */
         static 
 
-        int	fill_input_buffer(j_decompress_ptr cinfo);    
+        uchar fill_input_buffer(j_decompress_ptr cinfo);    
         static void skip_input_data(j_decompress_ptr cinfo, long count);
         static void term_source(j_decompress_ptr cinfo);
